+++ src/main/java/org/osgl/util/CollectorIterator.java
@@ -1,0 +1,80 @@
+package org.osgl.util;
+
+/*-
+ * #%L
+ * Java Tool
+ * %%
+ * Copyright (C) 2014 - 2017 OSGL (Open Source General Library)
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import org.osgl.$;
+
+import java.util.Iterator;
+
+class CollectorIterator<T, R> implements Iterator<R> {
+
+    private Iterator<? extends T> data;
+    private String path;
+
+    CollectorIterator
+    (Iterator<? extends T> itr, String path) {
+        this.data = $.requireNotNull(itr);
+        this.path = S.requireNotBlank(path);
+    }
+
+    protected Iterator<? extends T> data() {
+        return data;
+    }
+
+    @Override
+    public boolean hasNext() {
+        return data.hasNext();
+    }
+
+    @Override
+    public R next() {
+        return $.getProperty(data.next(), path);
+    }
+
+    @Override
+    public void remove() {
+        E.unsupport();
+    }
+
+    @Override
+    public int hashCode() {
+        return $.hc(data, path);
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == this) {
+            return true;
+        }
+        if (obj instanceof CollectorIterator) {
+            CollectorIterator that = (CollectorIterator)obj;
+            return $.eq(that.data, data) && $.eq(that.path, path);
+        }
+        return false;
+    }
+
+    @Override
+    public String toString() {
+        StringBuilder sb = new StringBuilder("CollectorIterator\npath[");
+        sb.append(path).append("]\nbuf[\n").append(data).append("\n]");
+        return sb.toString();
+    }
+}